# MAINT: speed up _block by avoiding a recursive closure