@eric-wieser 2018-09-20
This mostly just undoes the speed regression caused by #11910, I think