@eric-wieser 2018-09-20
@mattip, thoughts? I'm not sure how I feel about micro-optimizations like this that come at varying clarity costs.