@mattip 2018-09-20
It seems good to me, removing that recursive closure is +1 for clarity, and other than that it just seems to juggle the same functions that were called before as far as I am concerned.  I am personally not a big advocate of one-line functions, they break up the flow for me, so I would prefer to inline `_atleast_nd` (the comment should remain in any case), but not critical. Maybe someone will find some use for it in the future.